add: config 수정
