+++ config.yaml
@@ -5,12 +5,14 @@ train:
   model_name: 허깅페이스 모델명
   epoch: 100
   batch_size: 32
-  LR: 0.00001
-  LR_base: 20 # scheduler를 쓴다면, LR에 대한 min_lr을 적어주세요(20 == LR/20)
-  LR_max: 5 # scheduler를 쓴다면, LR에 대한 max_lr을 적어주세요(5 == LR/5)
-  LR_step_up: 10 # scheduler를 쓴다면, warmup epoch를 적어주세요
-  LR_step_down: 20 # scheduler를 쓴다면, cooldown epoch를 적어주세요 (단, up+down은 epoch과 동일해야 함)
-  lossF: 로스 함수명
+  LR:    #LambdaLR, StepLR, CyclicLR, ExponentialLR, WarmupConstantLR 
+    name: 스케줄러명
+    lr: 0.000001
+    base: 20 # CyclicLR를 쓴다면, LR에 대한 min_lr을 적어주세요(20 == LR/20)
+    max: 5 # CyclicLR를 쓴다면, LR에 대한 max_lr을 적어주세요(5 == LR/5)
+    step_up: 5 # CyclicLR를 쓴다면, warmup epoch를 적어주세요
+    step_down: 5 # CyclicLR를 쓴다면, cooldown epoch를 적어주세요 (단, up+down은 epoch과 동일해야 함)
+    warmupconstantLR_step: 3
   optim: 옵티마이저 함수명
   shuffle: True
   token_max_len: 100
